feature/polymorphic_testing - refactor
